[TASK] Link local resources (#161)
Local resources are linked instead of the ones from the Python-based
theme. This is necessary for the next steps, for example, adjust the
styles and JavaScripts to our needs. The formerly "_assets" folder was
renamed to "_resources" to match the folder in the theme package.
Additionally, the resources can now be copied recursively.

The styles for code highlighting are now embedded directly into the
codeblock.css file. This file was not linked before, this is now added.
Note: The code block has now a dark background color (as for the
Python-based rendering). This will be adjusted in a follow-up.